SelectAll for Checkboxgroupinput
Created a Select All Check Box Group Module2 that is a combination of checkboxgroup input and action buttons for select all, unselct all, and possibly select defaulted. Has arguments for column width and selected.
Incorporated the selectAll into the export tab
Changed the IDs of modules to not have any spaces. Spaces cause updatecheckboxgroupinput not to work (most likely the same for other update widget functions).